@@ -2,10 +2,13 @@ import { Injectable } from '@angular/core';
 import { CanActivate, CanActivateChild, CanLoad ,ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/first';
+import { first } from "rxjs/operators";
 
 import { AuthenticationService } from "./authentication.service";
 
+/**
+ * Guard that will be use in route if the current user is valid or not
+ */
 @Injectable()
 export class AuthenticationGuard implements CanActivate, CanActivateChild, CanLoad {
 
@@ -15,19 +18,22 @@ export class AuthenticationGuard implements CanActivate, CanActivateChild, CanLo
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean> {
 
-    return this._service.authenticateGuard().first();
+    return this._service.authenticateGuard()
+           .pipe( first() );
 
   }
 
   canActivateChild(route : ActivatedRouteSnapshot, state : RouterStateSnapshot) : Observable<boolean> {
 
-    return this._service.authenticateGuard().first();
+    return this._service.authenticateGuard()
+           .pipe( first() );
 
   }
 
   canLoad() : Observable<boolean> {
 
-    return this._service.authenticateGuard().first();
+    return this._service.authenticateGuard()
+           .pipe( first() );
 
   }
 
