@@ -4,9 +4,8 @@ import {  MatTabChangeEvent } from "@angular/material";
 
 import { Store } from "@ngrx/store";
 import { Observable } from "rxjs/Observable";
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/take';
+import { of } from "rxjs/observable/of";
+import { switchMap, take } from "rxjs/operators";
 
 import * as empActions from './actions/employee-details.actions';
 import * as personalActions from './employee-personal-information/actions/employee-personal.actions';
@@ -15,10 +14,13 @@ import * as fromRoot    from './reducers';
 import { EmployeePersonal } from "./employee-personal-information/models/employee-personal.models";
 import { EmployeeDetailsService } from "./employee-details.service";
 
+import { fadeAnimation } from "./../../animations/fade.animations";
+
 @Component({
   selector: 'app-employee-details',
   templateUrl: './employee-details.component.html',
   styleUrls: ['./employee-details.component.scss'],
+  animations:[fadeAnimation],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class EmployeeDetailsComponent implements OnInit {
@@ -37,8 +39,10 @@ export class EmployeeDetailsComponent implements OnInit {
   ngOnInit() {
 
     this._route.paramMap
-    .switchMap((params : ParamMap) => { return Observable.of(+params.get('id')) })
-    .take(1)
+    .pipe(
+      switchMap((params : ParamMap) => { return of(+params.get('id')) }),
+      take(1)
+    )
     .subscribe((id) => this._store.dispatch(new empActions.GetEmployeeId(id)) );
 
     this.checkRouteTabSelected();
@@ -78,6 +82,9 @@ export class EmployeeDetailsComponent implements OnInit {
     else if(index == 8){
       this._router.navigate(['club'], { relativeTo : this._route });
     }
+    else if(index == 9){
+      this._router.navigate(['account/setting'], { relativeTo : this._route });
+    }
 
   }
 
@@ -85,13 +92,25 @@ export class EmployeeDetailsComponent implements OnInit {
   checkRouteTabSelected(){
 
     this._route.firstChild.data
-    .take(1)
+    .pipe(
+      take(1)
+    )
     .subscribe((data) => {
 
      this.currentIndex = this._service.selectedRouteTab(data);
         
     });
 
+  }
+
+
+  /**
+   * This for animating every routes with a data 'animation'
+   * @param outlet 
+   */
+  prepRouteState(outlet : any){
+    return outlet.isActivated ? outlet.activatedRoute : '';
+    //return outlet.activatedRouteData.animation;
   }
 
 
