@@ -3,16 +3,8 @@ import { Injectable } from '@angular/core';
 import { Action, Store } from "@ngrx/store";
 import { Effect, Actions } from "@ngrx/effects";
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/withLatestFrom';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/distinctUntilChanged';
-import 'rxjs/add/operator/filter';
+import { of } from "rxjs/observable/of";
+import { map, switchMap, mergeMap, catchError, withLatestFrom, debounceTime, tap, distinctUntilChanged, filter } from "rxjs/operators";
 
 import * as projectActions from './../actions/project-table.actions';
 import * as fromRootProjects from './../reducers';
@@ -23,92 +15,126 @@ import { ProjectsService } from "./../projects.service";
 import { LoaderSpinnerService } from "./../../../main-content/services/loader-spinner/loader-spinner.service";
 import { ToastrService } from "./../../../main-content/services/toastr.service";
 
+import * as fromMasterData from './../../../master-data/reducers/master-data.reducers';
+import * as masterDataActions from './../../../master-data/actions/master-data.actions';
+
 @Injectable()
 export class ProjectsTableEffects {
 
-    constructor(private _service : ProjectsService, private _actions$ : Actions, private _store$ : Store<fromRootProjects.State>, private _loader : LoaderSpinnerService, private _toastr : ToastrService){}
+    constructor(private _service : ProjectsService, 
+                private _actions$ : Actions, 
+                private _store$ : Store<fromRootProjects.State>, 
+                private _masterStore$ : Store<fromMasterData.State>,
+                private _loader : LoaderSpinnerService, 
+                private _toastr : ToastrService){}
 
     @Effect()
         loadData$ : Observable<Action> = this._actions$
         .ofType(projectActions.LOAD)
-        .withLatestFrom(
-            this._store$.select(fromRootProjects.getCollectionPageIndex),
-            this._store$.select(fromRootProjects.getCollectionPageSize),
-            this._store$.select(fromRootProjects.getCollectionSortField),
-            this._store$.select(fromRootProjects.getCollectionSortDirection),
-            this._store$.select(fromRootProjects.getCollectionSearchQuery)
-        )
-        .switchMap( ([action, pageIndex, pageSize, sortField, sortDirection,searchQuery]) => {
-            return this._service.loadData(pageIndex,pageSize,sortField,sortDirection,searchQuery)
-            .map((response) => new projectActions.LoadSuccess(response.data,response.count) )
-            .catch(err => Observable.of(new projectActions.LoadError(err) ))
-
-        });
+        .pipe(
+            withLatestFrom(
+                this._store$.select(fromRootProjects.getCollectionPageIndex),
+                this._store$.select(fromRootProjects.getCollectionPageSize),
+                this._store$.select(fromRootProjects.getCollectionSortField),
+                this._store$.select(fromRootProjects.getCollectionSortDirection),
+                this._store$.select(fromRootProjects.getCollectionSearchQuery)
+            ),
+            switchMap( ([action, pageIndex, pageSize, sortField, sortDirection,searchQuery]) => {
+                return this._service.loadData(pageIndex,pageSize,sortField,sortDirection,searchQuery)
+                    .pipe(
+                        map((response) => new projectActions.LoadSuccess(response.data,response.count) ),
+                        catchError(err => of(new projectActions.LoadError(err) ))
+                    )    
+            })
+        );
 
     @Effect()
         search$ : Observable<Action> = this._actions$
         .ofType(projectActions.SEARCH)
-        .withLatestFrom(
-            this._store$.select(fromRootProjects.getCollectionPageIndex),
-            this._store$.select(fromRootProjects.getCollectionPageSize),
-            this._store$.select(fromRootProjects.getCollectionSortField),
-            this._store$.select(fromRootProjects.getCollectionSortDirection),
-            this._store$.select(fromRootProjects.getCollectionSearchQuery)
-        )
-        .debounceTime(300)
-        .distinctUntilChanged()
-        .switchMap( ([action, pageIndex, pageSize, sortField, sortDirection,searchQuery]) => {
-            return this._service.loadData(pageIndex,pageSize,sortField,sortDirection,searchQuery)
-            .map((response) => new projectActions.LoadSuccess(response.data,response.count) )
-            .catch(err => Observable.of(new projectActions.LoadError(err) ))
-
-        });
+        .pipe(
+            withLatestFrom(
+                this._store$.select(fromRootProjects.getCollectionPageIndex),
+                this._store$.select(fromRootProjects.getCollectionPageSize),
+                this._store$.select(fromRootProjects.getCollectionSortField),
+                this._store$.select(fromRootProjects.getCollectionSortDirection),
+                this._store$.select(fromRootProjects.getCollectionSearchQuery)
+            ),
+            debounceTime(300),
+            distinctUntilChanged(),
+            switchMap( ([action, pageIndex, pageSize, sortField, sortDirection,searchQuery]) => {
+                return this._service.loadData(pageIndex,pageSize,sortField,sortDirection,searchQuery)
+                    .pipe(
+                        map((response) => new projectActions.LoadSuccess(response.data,response.count) ),
+                        catchError(err => of(new projectActions.LoadError(err) ))
+                    )
+    
+            })
+        );
 
     @Effect()
         save$ : Observable<Action> = this._actions$
         .ofType<projectActions.SaveProject>(projectActions.SAVE_PROJECT)
-        .map( (action) => action.payload)
-        .switchMap((payload) => {
-
-            this._loader.openDialog();
-            if(payload.projectId == 0){
-            
-                return this._service.save(payload)
-                .map((response) =>  new projectActions.SaveSuccess(response.status) )
-                .catch((err) => Observable.of( new projectActions.LoadError(err) ))
-                .do(() => this._loader.closeDialog())
+        .pipe(
+            map( (action) => action.payload),
+            switchMap((payload) => {
+    
+                this._loader.openDialog();
+                if(payload.projectId == 0){
                 
-            }
-            else {
-
-                return this._service.update(payload)
-                .map((response) => new projectActions.SaveSuccess(response.status) )
-                .catch((err) => Observable.of( new projectActions.LoadError(err) ))
-                .do(() => this._loader.closeDialog())
-            }
-
-        })
-
+                    return this._service.save(payload)
+                        .pipe(
+                            mergeMap((response) => {
+                                return [
+                                    new projectActions.SaveSuccess(response.status),
+                                    new projectActions.Load(),
+                                    new masterDataActions.AddNewProject(response.createdData)         
+                                ];
+                            }),
+                            catchError((err) => of( new projectActions.LoadError(err) )),
+                            tap(() => this._loader.closeDialog())
+                        )
+                
+                }
+                else {
+    
+                    return this._service.update(payload)
+                        .pipe(
+                            mergeMap((response) => {
+                                return [
+                                    new projectActions.SaveSuccess(response.status),
+                                    new projectActions.UpdateSuccess({ id: payload.projectId, updatedData: payload  }),
+                                    new masterDataActions.UpdateProject({ id: payload.projectId, updatedData: payload  })
+                                ];
+                            }),
+                            catchError((err) => of( new projectActions.LoadError(err) )),
+                            tap(() => this._loader.closeDialog())
+                        )
+                }
+    
+            })    
+        );
+        
     @Effect()
-        saveSuccess$  = this._actions$
+        saveSuccess$ : Observable<Action>  = this._actions$
                        .ofType(projectActions.SAVE_SUCCESS)
-                       .do(() => { this._toastr.saveSuccess(); })
-                       .mergeMap(() => {
-                           return [
-                            new projectActions.Load(),
-                            new projectActions.ClearSelectProject()
-                           ];
-                       })
+                       .pipe(
+                            tap(() => { this._toastr.saveSuccess(); }),
+                            map(() =>  new projectActions.ClearSelectProject())
+                           
+                       );
                        
 
                        
 
     @Effect()
-        error$ = this._actions$
+        error$ : Observable<Action> = this._actions$
                  .ofType<projectActions.LoadError>(projectActions.LOAD_ERROR)
-                 .map((action) => action.payload)
-                 .do((payload) => { this._toastr.errorHandler(payload)})
-                 .filter( payload => payload.status == 422 )
-                 .map(() => new projectActions.ClearSelectProject() )
+                 .pipe(
+                    map((action) => action.payload),
+                    tap((payload) => { this._toastr.errorHandler(payload)}),
+                    filter( payload => payload.status == 422 ),
+                    map(() => new projectActions.ClearSelectProject() )
+                 );
                  
+            
 }
