@@ -28,7 +28,7 @@ import { environment } from "./../environments/environment";
 
 @NgModule({
   declarations: [
-    AppComponent,
+    AppComponent
   ],
   imports: [
     BrowserModule,
