@@ -38,6 +38,7 @@ export class PositionsFormComponent implements OnInit, OnDestroy {
 
         this.positionForm.setValue({
           positionId   : response.positionId,
+          positionCode : response.positionCode,
           positionName : response.positionName
         });
 
@@ -46,19 +47,22 @@ export class PositionsFormComponent implements OnInit, OnDestroy {
 
   }
   
-  get currentId() { return this.positionForm.get('positionId').value };
+
+  getCurrentId() {
+    return this.positionForm.get('positionId').value
+  }
 
 
   createForm() {
     this.positionForm = this._fb.group({
         positionId    :  [0,Validators.required],
-        positionName  :  [null,Validators.required]
+        positionCode  :  [null,[Validators.required, Validators.maxLength(20)]],
+        positionName  :  [null,[Validators.required, Validators.maxLength(150)]]
     });
   }
 
   submitForm(){
 
-    this._store$.dispatch( new positionActions.SelectPosition(this.positionForm.value) );
     this._store$.dispatch( new positionActions.SavePosition(this.positionForm.value) );
     this._dialogRef.close();
     
