@@ -1,16 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 
 import { Store } from "@ngrx/store";
 import { Observable } from 'rxjs/Observable';
 import { Subscription } from "rxjs/Subscription";
-
 
 @Component({
   selector: 'app-breadcrumb',
   templateUrl: './breadcrumb.component.html',
   styleUrls: ['./breadcrumb.component.scss']
 })
-export class BreadcrumbComponent implements OnInit {
+export class BreadcrumbComponent implements OnInit, OnDestroy {
   subscription : Subscription
   urlRouterState : Observable<any>;
   breadcrumbs : any[] = [];
@@ -20,6 +19,7 @@ export class BreadcrumbComponent implements OnInit {
   }
   
   ngOnInit(){
+
     this.subscription = this.urlRouterState
       .subscribe((response) => {
         if(response != undefined){
@@ -27,6 +27,12 @@ export class BreadcrumbComponent implements OnInit {
           this.breadcrumbs = str.split("/");
         }
       });
+     
+
+  }
+
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
   
 }
