@@ -1,7 +1,13 @@
 
+import { RouterLinks } from "./../models/router-links.model";
 
+// A Collection of links per user role
 
-export const sidenavMetaData  = [
+export const defaultLinks : RouterLinks[] = [
+    { title : '404 NOT FOUND', isHeader: true }
+];
+
+export const superAdminLinks : RouterLinks[]  = [
     { title : 'Master Data' , isHeader : true},
     { title : 'Projects' , link : 'projects', icon : 'table', isHeader : false },
     { title : 'Positions' , link : 'positions', icon : 'table', isHeader : false },
@@ -14,10 +20,12 @@ export const sidenavMetaData  = [
 
 
 
-export const sidenavMetaData2  = [
-    { title : 'Accounting' , isHeader : true},
-    { title : 'Dashboard 1' , link : '', icon : 'menu', isHeader : false },
-    { title : 'Dashboard 2' , link : '', icon : 'menu', isHeader : false },
-    { title : 'Dashboard 3' , link : '', icon : 'menu', isHeader : false },
-    { title : 'Dashboard 4' , link : '', icon : 'menu', isHeader : false },
+export const projectHrLinks : RouterLinks[] = [
+    { title : 'Master Data' , isHeader : true},
+    { title : 'Positions' , link : 'positions', icon : 'table', isHeader : false },
+    { title : 'Employment Status' , link : 'employment/status', icon : 'table', isHeader : false },
+    { title : 'Employee Status' , link : 'employee/status', icon : 'table', isHeader : false },
+    { title : 'Human Resources' , isHeader : true},
+    { title : 'Register Employee' , link : 'employee/register', icon : 'book', isHeader : false },
+    { title : 'Employee List' , link : 'employee/list', icon : 'people', isHeader : false },
 ];
