@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpParams } from "@angular/common/http";
-
+import { HttpClient } from "@angular/common/http";
 
 import { Projects } from "./../features/projects/models/projects.model";
 import { Position } from "./../features/positions/models/positions.model";
@@ -36,19 +35,19 @@ export class MasterDataService {
   restEndPoint : string = environment.endPoint;
   constructor(private _http : HttpClient) { }
 
-  getProjects() {
+  getAllProjects() {
     return this._http.get<ProjectResponse>(`${this.restEndPoint}/projects/all`);
   }
 
-  getPositions() {
+  getAllPositions() {
     return this._http.get<PositionResponse>(`${this.restEndPoint}/positions/all`);
   }
 
-  getEmploymentStatus() {
+  getAllEmploymentStatus() {
     return this._http.get<EmploymentStatusResponse>(`${this.restEndPoint}/employmentStatus/all`);
   }
 
-  getEmployeeStatus() {
+  getAllEmployeeStatus() {
     return this._http.get<EmployeeStatusResponse>(`${this.restEndPoint}/employeeStatus/all`);
   }
 
