@@ -1,5 +1,9 @@
 
-import { ActionReducerMap , ActionReducer, MetaReducer } from '@ngrx/store';
+/**
+ * This is the root reducers of your entire application
+ * The feature reducers state extends this reducers state
+ */
+import { ActionReducerMap , ActionReducer, MetaReducer, Action } from '@ngrx/store';
 import { environment } from './../../environments/environment';
 
 import { RouterStateUrl } from "./../shared/utils";
@@ -17,10 +21,11 @@ import { storeFreeze } from "ngrx-store-freeze";
  * the state of the reducer plus any selector functions. The `* as`
  * notation packages up all of the exports into a single object.
  */
-import * as fromMainContent from './../main-content/reducers/main-content.reducers';
-import * as fromRouterLinks from './../main-content/reducers/router-links.reducers';
-import * as fromMasterData  from './../master-data/reducers/master-data.reducers';
+import * as fromMainContent   from './../main-content/reducers/main-content.reducers';
+import * as fromRouterLinks   from './../main-content/reducers/router-links.reducers';
+import * as fromMasterData    from './../master-data/reducers/master-data.reducers';
 import * as fromLoginPresence from './../features/login/reducers/login.reducers';
+
 /**
  * As mentioned, we treat each reducer like a table in a database. This means
  * our top level state interface is just a map of keys to inner state types.
@@ -42,12 +47,26 @@ export interface State {
  */
 
 export const reducers   :   ActionReducerMap<State> = {
-    mainContent     : fromMainContent.reducer,
-    routerLinks     : fromRouterLinks.reducer,
-    masterData      : fromMasterData.reducer,
-    loginPresence   : fromLoginPresence.reducer,
-    routerReducer   : fromRouter.routerReducer
+    mainContent         : fromMainContent.reducer,
+    routerLinks         : fromRouterLinks.reducer,
+    masterData          : fromMasterData.reducer,
+    loginPresence       : fromLoginPresence.reducer,
+    routerReducer       : fromRouter.routerReducer
   };
+
+
+export function clearState(reducer : ActionReducer<State>) : ActionReducer<State> {
+
+  return function (state : State, action : Action) : State {
+
+    if(action.type === '[LOGIN] Logout'){
+      state = undefined;
+    }
+
+    return reducer(state,action);
+  }
+
+} 
 
 
 /**
@@ -56,7 +75,7 @@ export const reducers   :   ActionReducerMap<State> = {
  * that will be composed to form the root meta-reducer.
  */
 export const metaReducers: MetaReducer<State>[] = !environment.production
-? [storeFreeze]
-: [];
+? [storeFreeze,clearState]
+: [clearState];
 
 
